feat(explorer): use child routes for node show/edit + fix layout issues
