+++ .editorconfig
@@ -15,6 +15,18 @@ insert_final_newline = true
 indent_style = tab
 indent_size = 4
 
+[*.js]
+indent_style = space
+indent_size = 4
+
+[*.jsx]
+indent_style = space
+indent_size = 4
+
+[*.styl]
+indent_style = space
+indent_size = 2
+
 [*.json]
 insert_final_newline = false
