+++ explorer/src/components/nodes/NodeEdit.jsx
@@ -1,9 +1,9 @@
-import React, { Component, PropTypes }   from 'react';
-import { connect }                       from 'react-redux';
-import { Link }                          from 'react-router';
-import { FormattedMessage }              from 'react-intl';
-import NodeForm                          from './NodeForm.jsx';
-import { fetchNodeIfNeeded, updateNode } from '../../actions';
+import React, { Component, PropTypes } from 'react';
+import { connect }                     from 'react-redux';
+import { Link }                        from 'react-router';
+import { FormattedMessage }            from 'react-intl';
+import NodeForm                        from './NodeForm.jsx';
+import { updateNode }                  from '../../actions';
 
 const assign = Object.assign || require('object.assign');
 
@@ -12,26 +12,15 @@ class NodeEdit extends Component {
     static displayName = 'NodeEdit';
 
     static propTypes = {
-        node:                      PropTypes.object,
-        isFetching:                PropTypes.bool.isRequired,
-        dispatchFetchNodeIfNeeded: PropTypes.func.isRequired,
-        dispatchNodeUpdate:        PropTypes.func.isRequired,
-        routeParams:               PropTypes.object.isRequired
+        node:               PropTypes.object,
+        isFetching:         PropTypes.bool.isRequired,
+        dispatchNodeUpdate: PropTypes.func.isRequired
     };
 
     constructor(props) {
         super(props);
 
         this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    fetchNode() {
-        const { dispatchFetchNodeIfNeeded, routeParams } = this.props;
-        dispatchFetchNodeIfNeeded(routeParams.node_uuid);
-    }
-
-    componentDidMount() {
-        this.fetchNode();
     }
 
     handleSubmit(data) {
@@ -49,11 +38,13 @@ class NodeEdit extends Component {
         }
 
         return (
-            <div>
+            <div className="node-main">
                 <h1 className="panel-title">
                     <FormattedMessage id="node.edit.title" values={{ name: node.name }}/>
                 </h1>
-                <NodeForm onSubmit={this.handleSubmit} initialValues={node}/>
+                <div className="panel-body">
+                    <NodeForm onSubmit={this.handleSubmit} initialValues={node}/>
+                </div>
             </div>
         );
     }
@@ -73,8 +64,7 @@ const mapStateToProps = ({ nodes, nodesByUuid }) => {
 };
 
 const mapDispatchToProps = dispatch => ({
-    dispatchFetchNodeIfNeeded: (nodeUuid) => dispatch(fetchNodeIfNeeded(nodeUuid)),
-    dispatchNodeUpdate:        (node)     => dispatch(updateNode(node))
+    dispatchNodeUpdate: (node) => dispatch(updateNode(node))
 });
