+++ explorer/src/components/nodes/NodeRevisions.jsx
@@ -1,4 +1,5 @@
 import React, { Component, PropTypes } from 'react';
+import { connect }                     from 'react-redux';
 import NodeRevisionsItem               from './NodeRevisionsItem.jsx';
 import { fetchNodeRevisionsIfNeeded }  from '../../actions';
 
@@ -7,18 +8,15 @@ class NodeRevisions extends Component {
     static displayName = 'NodeRevisions';
 
     static propTypes = {
-        uuid:      PropTypes.string.isRequired,
-        revisions: PropTypes.array.isRequired,
-        dispatch:  PropTypes.func.isRequired
+        uuid:           PropTypes.string.isRequired,
+        isFetching:     PropTypes.bool.isRequired,
+        revisions:      PropTypes.array.isRequired,
+        fetchRevisions: PropTypes.func.isRequired
     };
 
-    fetchRevisions() {
-        const { dispatch, uuid } = this.props;
-        dispatch(fetchNodeRevisionsIfNeeded(uuid));
-    }
-
     componentDidMount() {
-        this.fetchRevisions();
+        const { fetchRevisions, uuid } = this.props;
+        fetchRevisions(uuid);
     }
 
     render() {
@@ -38,5 +36,27 @@ class NodeRevisions extends Component {
     }
 }
 
+const mapStateToProps = ({ nodes, nodesRevisionsByUuid }) => {
+    let revisions = {
+        items:      [],
+        isFetching: true
+    };
+    if (nodesRevisionsByUuid[nodes.currentUuid]) {
+        revisions = nodesRevisionsByUuid[nodes.currentUuid];
+    }
 
-export default NodeRevisions;
+    return {
+        isFetching: revisions.isFetching,
+        revisions:  revisions.items
+    };
+};
+
+const mapDispatchToProps = dispatch => ({
+    fetchRevisions: (nodeUuid) => dispatch(fetchNodeRevisionsIfNeeded(nodeUuid))
+});
+
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(NodeRevisions);
