+++ explorer/src/index.html
@@ -3,7 +3,7 @@
     <meta http-equiv="Content-type" content="text/html; charset=utf-8"/>
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
     <title>APPUI | gonode client</title>
-    <link href='https://fonts.googleapis.com/css?family=Raleway:400,500,300,700,900' rel='stylesheet' type='text/css'>
+    <link href='https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700,800|Montserrat:400,700' rel='stylesheet' type='text/css'>
 </head>
 <body>
     <div id="app"></div>
