+++ explorer/src/styles/nodes-list.styl
@@ -1,4 +1,11 @@
 .nodes-list
+  position absolute
+  top ($header-height + $pager-height)
+  left   0
+  right  0
+  bottom 0
+  overflow-y auto
+
   &_item
     display: block
     padding: 9px 20px
