+++ explorer/src/styles/pager.styl
@@ -1,11 +1,14 @@
 .pager
+  position absolute
+  top $header-height
+  right 0
+  left  0
   background: $color-pager-background
   height: $pager-height
   line-height: $pager-height
   font-size: 14px
   text-align: center
   padding: 0 $pager-nav-width
-  position: relative
 
   &_page
     display: inline-block
