+++ explorer/src/styles/settings/dimensions.styl
@@ -1,4 +1,4 @@
-$navigation-width = 240px
+$navigation-width        = 240px
 $nodes-list-marker-width = 4px
 
 $login-width         = 300px
@@ -20,8 +20,8 @@ $node-revisions-stroke-width = 2px
 // ----------------------------------------
 // Pager
 // ----------------------------------------
-$pager-height = 36px
-$pager-nav-width = 36px
+$pager-height    = 52px
+$pager-nav-width = 46px
 
 
 // ----------------------------------------
